+++ CHANGELOG.rst
@@ -1,7 +1,14 @@
+0.1.2 - 30-Nov-2017
+-----------------
+
+* Fix ``ip2geotools.databases.commercial.DbIpWeb`` because of new webpage layout
+* "IP address not found" error can be recognized in ``ip2geotools.databases.commercial.SkyhookContextAcceleratorIp``
+* Custom exceptions can be formatted in ``ip2geotools.errors``
+
 0.1.1 - 01-Nov-2017
 -----------------
 
-* Fix installation from PyPi using pip
+* Fix installation from PyPi using ``pip``
 
 0.1 - 01-Nov-2017
 -----------------